+++ terraform_gcp_infra/environments/prod/proj-default-templet/10-network/main.tf
@@ -62,3 +62,27 @@ module "net" {
 
   firewall_rules = var.firewall_rules
 }
+
+locals {
+  private_service_connection_name = length(trimspace(var.private_service_connection_name)) > 0 ? var.private_service_connection_name : "${module.naming.vpc_name}-psc"
+}
+
+resource "google_compute_global_address" "private_service_connect" {
+  count        = var.enable_private_service_connection ? 1 : 0
+  name         = local.private_service_connection_name
+  project      = var.project_id
+  purpose      = "VPC_PEERING"
+  address_type = "INTERNAL"
+
+  prefix_length = var.private_service_connection_prefix_length
+  network       = module.net.vpc_self_link
+}
+
+resource "google_service_networking_connection" "private_vpc_connection" {
+  count                   = var.enable_private_service_connection ? 1 : 0
+  network                 = module.net.vpc_self_link
+  service                 = "services/servicenetworking.googleapis.com"
+  reserved_peering_ranges = [google_compute_global_address.private_service_connect[0].name]
+
+  depends_on = [google_compute_global_address.private_service_connect]
+}